requirements.txt
absl-py==2.1.0
anyio==4.3.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
audioread==3.0.1
Babel==2.15.0
beautifulsoup4==4.12.3
bleach==6.1.0
cachetools==5.5.2
certifi==2024.2.2
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cloudpickle==3.0.0
comm==0.2.2
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
docker-pycreds==0.4.0
durationpy==0.9
exceptiongroup==1.1.3
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.1
flatbuffers==24.3.25
fonttools==4.51.0
fqdn==1.5.1
fsspec==2024.5.0
gast==0.5.4
gdown==5.1.0
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.38.0
google-pasta==0.2.0
grpcio==1.62.1
gym==0.26.2
gym-notices==0.0.8
h11==0.14.0
h5py==3.11.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.23.0
idna==3.4
iniconfig==2.0.0
ipykernel==6.29.3
ipython==8.22.2
ipython-genutils==0.2.0
ipywidgets==8.1.2
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.20
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.4.0
jupyter-events==0.9.0
jupyter-lsp==2.2.4
jupyter_client==8.6.0
jupyter_core==5.7.2
jupyter_server==2.13.0
jupyter_server_terminals==0.5.2
jupyterlab==4.1.3
jupyterlab-pygments==0.2.2
jupyterlab_server==2.25.3
jupyterlab_widgets==3.0.10
keras==3.3.3
kiwisolver==1.4.5
kubernetes==32.0.1
lazy_loader==0.4
libclang==14.0.1
librosa==0.10.1
llvmlite==0.41.1
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.4
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.3.2
mpmath==1.3.0
msgpack==1.0.4
namex==0.0.8
nbclient==0.8.0
nbconvert==7.16.2
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.3
notebook==7.1.1
notebook_shim==0.2.4
numba==0.58.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opt-einsum==3.3.0
optree==0.10.0
overrides==7.7.0
packaging==23.2
pandas==2.2.1
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pillow==10.2.0
platformdirs==4.2.0
pluggy==1.5.0
pooch==1.8.1
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.2
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-spy==0.3.14
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.21
Pygments==2.18.0
pyparsing==3.1.1
PySocks==1.7.1
pytest==8.2.0
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
python-Levenshtein==0.12.2
pytorch-ranger==0.1.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.3.2
QtPy==2.2.0
referencing==0.34.0
regex==2023.8.8
requests==2.31.0
requests-oauthlib==2.0.0
resampy==0.2.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.10.6
rsa==4.9
safetensors==0.4.1
scikit-learn==1.4.2
scipy==1.12.0
Send2Trash==1.8.2
sentencepiece==0.1.99
sentry-sdk==2.2.1
setproctitle==1.3.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.5
stack-data==0.6.3
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.0
tensorboardX==2.6.2.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.4.0
terminado==0.18.0
threadpoolctl==3.5.0
tinycss2==1.2.1
tokenizers==0.19.1
tomli==2.0.1
torch==2.3.0
torch-optimizer==0.3.0
torchvision==0.18.0
tornado==6.3.3
tqdm==4.66.4
traitlets==5.14.2
transformers==4.40.2
triton==2.3.0
types-python-dateutil==2.8.19.20240106
typing_extensions==4.11.0
tzdata==2024.1
uri-template==1.3.0
urllib3==2.1.0
wandb==0.16.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
widgetsnbextension==4.0.10
wrapt==1.16.0
